uploaded example of collapsible box
The styles.css file inside the collapsible folder contains all the css needed for the collapsible box stuff. One thing I didn't include was symbols indicating whether the box was open or closed; that's not too hard to add if necessary. The collapsible.html itself just has an example of the box and script needed.